+++ CHANGELOG.md
@@ -16,7 +16,7 @@ This changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.
 ### Tests
 
 ### .github
-
+- Added GitHub Actions workflow to remind draft PR authors to mark ready for review after pushing changes. (#1722)
 
 
 ## [0.2.0] - 2026-11-02